Pri-Med-Final-master/requirements.txt
Flask
werkzeug
tensorflow
tenseal
torch
torchvision
Pillow
